Remove branch env var, use HEAD
I think we can rely on the GitHub checkout action to set the HEAD to be the commit that the user wants to deploy.

Even if they want to deploy a different branch to the one that triggered the action, they can still do this by running the checkout action with the right arguments.

This removes the git checkout line from this script too, so I think it’ll make it faster?